I did a bad thing. I changed multiple features in one commit. In one branch.
List of not completely inclusive changes:

Add repositories User and a test repository.

Started a AccountController Test file. It's admittedly difficult to do anything involving the identity, because OWin requires additional configuration.

Add the Orders table and class to the model, as well as auto generated the classes (via Diagram save).